third_party/WebKit/LayoutTests/external/wpt/html/browsers/history/the-location-interface/scripted_click_location_assign_during_load-expected.txt

This is a testharness.js-based test.
FAIL location.assign with click during loadassert_equals: expected 2 but got 1
Harness: the test ran to completion.
